Implemented deletion of items
This commit adds in a basic deletion mode. It allows you to press a button
to enter "Delete Mode." While in this mode, any item clicked will be
"deleted." The items are not actually removed yet, and the action can be
undone. When finished, any items chosen will be permanently deleted.