Added graph library, fixed session delete bug
Added the graphview library, allowing for the plotting of session data
points. Currently only a basic graph is implemented.

Fixed a bug that would crash the app if you spam clicked when deleting
items. getAdapterPosition() could return -1 if a click was received in
empty area, which is possible when the recyclerview is updating positions
of items.