Added Constants class, replaced hardcoded items
Added in the Constants class so that we can define things like the ID for
an Extra in a single place.
Replaced all the hardcoded items with a value in Constants.

Switching to the SessionsViewActivity will now load up notes for the
currently selected note instead of always just being C#.