Implemented dialog to allow session naming
Gave the toggleRecordButton in MainActivity an onClickListener in
MainActivity.

Removed AssociatedMP3 from Session object.

Removed AssociatedMP3 column from Database. This has caused the database
version to increase. With the current implementation, an older database
version will delete all its existing data before upgrading the database to
the new version.

More tweaks to the graph in ItemActivity.

ToggleRecord has been removed as it is not used or needed.

The functions associated with generateRandomSession have been removed from
Util.

The noteSelectButton and historyViewButton are now disabled and hidden
when recording a session.