Database (#8)
* Added Constants class, replaced hardcoded items

Added in the Constants class so that we can define things like the ID for
an Extra in a single place.
Replaced all the hardcoded items with a value in Constants.

Switching to the SessionsViewActivity will now load up notes for the
currently selected note instead of always just being C#.

* Miscellaneous changes

Values in the Note enum have been reordered to properly reflect the scale
of notes.

The graph in ItemActivity has been fixed and now shows data correctly.

Redundant and commented out items have been removed from
activity_note_selet.xml.

activity_history_view.xml has been removed because it was not used
anywhere.

Items in activity_item.xml have been organized into a LinearLayout.

colorPrimaryDark has been updated to be the same color as the top of the
gradient, giving it the illusion of being transparent.

* Implemented dialog to allow session naming

Gave the toggleRecordButton in MainActivity an onClickListener in
MainActivity.

Removed AssociatedMP3 from Session object.

Removed AssociatedMP3 column from Database. This has caused the database
version to increase. With the current implementation, an older database
version will delete all its existing data before upgrading the database to
the new version.

More tweaks to the graph in ItemActivity.

ToggleRecord has been removed as it is not used or needed.

The functions associated with generateRandomSession have been removed from
Util.

The noteSelectButton and historyViewButton are now disabled and hidden
when recording a session.